+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/action/BrowserAction.kt
@@ -963,6 +963,11 @@ sealed class DownloadAction : BrowserAction() {
     data class UpdateDownloadAction(val download: DownloadState) : DownloadAction()
 
     /**
+     * Mark the download notification of the provided [downloadId] as removed from the status bar.
+     */
+    data class DismissDownloadNotificationAction(val downloadId: String) : DownloadAction()
+
+    /**
      * Restores the [BrowserState.downloads] state from the storage.
      */
     object RestoreDownloadsStateAction : DownloadAction()